Conflicts:
	debian/changelog
	debian/control
	debian/copyright
	debian/gdal-symbols.pl
	debian/libgdal1h.lintian-overrides
	debian/libgdal1h.symbols.amd64
	debian/libgdal1h.symbols.armel
	debian/libgdal1h.symbols.armhf
	debian/libgdal1h.symbols.i386
	debian/libgdal1h.symbols.ia64
	debian/libgdal1h.symbols.mips
	debian/libgdal1h.symbols.mipsel
	debian/libgdal1h.symbols.powerpc
	debian/libgdal1h.symbols.s390x
	debian/libgdal1h.symbols.sparc
	debian/patches/accomodate-typo
	debian/patches/algoritm-typo
	debian/patches/allows-typo
	debian/patches/bandwith-typo
	debian/patches/consistant-typo
	debian/patches/developement-typo
	debian/patches/explicitely-typo
	debian/patches/hurd-i386-pathmax
	debian/patches/inconsistant-typo
	debian/patches/initalize-typo
	debian/patches/paramters-typo
	debian/patches/perl-vendor
	debian/patches/ressource-typo
	debian/patches/series
	debian/patches/supress-typo
	debian/patches/transfered-typo
	debian/patches/transfering-typo
	debian/patches/unkown-typo
	debian/patches/usefull-typo

+++ ogr/ogrsf_frmts/shape/shptree.c
@@ -7,6 +7,7 @@
  *
  ******************************************************************************
  * Copyright (c) 1999, Frank Warmerdam
+ * Copyright (c) 2012, Even Rouault <even dot rouault at mines-paris dot org>
  *
  * This software is available under the following "MIT Style" license,
  * or at the option of the licensee under the LGPL (see LICENSE.LGPL).  This
@@ -910,13 +911,13 @@ SHPSearchDiskTreeNode( SHPTreeDiskHandle hDiskTree, double *padfBoundsMin, doubl
 /* -------------------------------------------------------------------- */
     if(numshapes > 0) 
     {
-        if( *pnResultCount + numshapes > *pnBufferMax )
+        if( (int)(*pnResultCount + numshapes) > *pnBufferMax )
         {
             int* pNewBuffer;
 
             *pnBufferMax = (*pnResultCount + numshapes + 100) * 5 / 4;
 
-            if( *pnBufferMax > INT_MAX / sizeof(int) )
+            if( (unsigned int)*pnBufferMax > INT_MAX / sizeof(int) )
                 *pnBufferMax = *pnResultCount + numshapes;
 
             pNewBuffer = (int *)
